Progress-1-Simple-LMS-Docker-Django-Foundation/requirements.txt
Django>=4.0
psycopg2-binary
gunicorn
django-ninja
django-ninja-jwt
